requirements.txt
python-telegram-bot~=21.6
azure-functions~=1.21.3
config~=0.5.1
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.13.3
utils~=1.0.2
flask~=3.0.3